+++ tests/payment/models/Payment - Credit card - Error Finishing Transaction.model.js
@@ -10,7 +10,9 @@ import {
 } from '../../../utils/shipping-actions'
 import {
   completePurchase,
+  getIframeBody,
   payWithCreditCard,
+  queryIframe,
 } from '../../../utils/payment-actions'
 import { ACCOUNT_NAMES } from '../../../utils/constants'
 
@@ -59,6 +61,76 @@ export default function test(account) {
       )
       cy.get('.payment-unauthorized-button').click()
       cy.get('.payment-unauthorized-modal').should('not.be.visible')
+    })
+
+    it('should be able to update billing address after error', () => {
+      const email = getRandomEmail()
+
+      setup({ skus: ['289'], account })
+      fillEmail(email)
+      fillProfile({
+        lastName: 'Denied',
+      })
+      fillShippingInformation(account)
+      goToPayment()
+      payWithCreditCard()
+      completePurchase()
+
+      cy.get('.payment-unauthorized-modal', { timeout: 120000 }).should(
+        'be.visible'
+      )
+      cy.get('.payment-unauthorized-button').click()
+      cy.get('.payment-unauthorized-modal').should('not.be.visible')
+
+      queryIframe($iframe => {
+        const body = getIframeBody($iframe)
+
+        cy.wrap(body)
+          .find(`#address-toggle-0`)
+          .click()
+
+        cy.wrap(body)
+          .contains('Endereço de cobrança')
+          .should('be.visible')
+      })
+    })
+
+    it('should validate form fields after error', () => {
+      const email = getRandomEmail()
+
+      setup({ skus: ['289'], account })
+      fillEmail(email)
+      fillProfile({
+        lastName: 'Denied',
+      })
+      fillShippingInformation(account)
+      goToPayment()
+      payWithCreditCard()
+      completePurchase()
+
+      cy.get('.payment-unauthorized-modal', { timeout: 120000 }).should(
+        'be.visible'
+      )
+      cy.get('.payment-unauthorized-button').click()
+      cy.get('.payment-unauthorized-modal').should('not.be.visible')
+
+      queryIframe($iframe => {
+        const body = getIframeBody($iframe)
+
+        cy.wrap(body)
+          .find('#creditCardpayment-card-0Code')
+          .clear()
+      })
+
+      completePurchase()
+
+      queryIframe($iframe => {
+        const body = getIframeBody($iframe)
+
+        cy.wrap(body)
+          .contains('Campo obrigatório')
+          .should('be.visible')
+      })
     })
   })
 }